pyproject.toml
[project]
name = "django-wysiwyg"
version = "0.1.0"
description = "django + wysiwyg をやってみる"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Django==5.1.2",
    "django-tinymce==4.1.0",
    "django-environ==0.12.0",
    "pillow==11.1.0",
]

[tool.uv]
link-mode = "copy"
dev-dependencies = [
    "ruff==0.9.2",
]
